backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
sqlalchemy==2.0.29
pydantic==2.6.4
pydantic-settings==2.2.1
requests==2.31.0
dateparser==1.1.4
google-api-python-client==2.119.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
oauth2client==4.1.3
spacy==3.6.1
python-telegram-bot==20.7